Working on slope search for p-adic fields
